+++ resources/js/core/layouts/Default.vue
@@ -11,6 +11,7 @@
                 { 'is-collapsed' : !menu.isExpanded }
             ]" v-if="menu.isVisible"/>
             <section :class="['main-content', menu.isExpanded ? 'is-expanded' : 'is-collapsed' ]">
+                <history v-if="history"/>
                 <div class="container is-fluid page-content is-marginless">
                     <page-header :title="$route.meta.title"/>
                     <router v-if="isInitialised"/>
@@ -26,24 +27,26 @@
 
 <script>
 
-import { mapState, mapMutations, mapActions } from 'vuex';
+import { mapState, mapGetters, mapMutations, mapActions } from 'vuex';
 import Navbar from '../structure/navbar/Navbar.vue';
 import Sidebar from '../structure/menu/Sidebar.vue';
 import Settings from '../structure/settings/Settings.vue';
 import AppFooter from '../structure/AppFooter.vue';
 import Router from '../Router.vue';
+import History from '../structure/History.vue';
 import PageHeader from '../structure/PageHeader.vue';
 
 export default {
     name: 'Default',
 
     components: {
-        Navbar, Sidebar, Settings, AppFooter, Router, PageHeader,
+        Navbar, Sidebar, Settings, AppFooter, Router, History, PageHeader,
     },
 
     computed: {
         ...mapState(['meta', 'isInitialised']),
         ...mapState('layout', ['lightsOff', 'isTablet', 'isMobile', 'menu', 'settingsBar']),
+        ...mapGetters('preferences', ['history']),
         lightsOn() {
             return !this.lightsOff;
         },