+++ resources/js/middleware/after.js
@@ -1,8 +1,10 @@
 import documentTitle from './after/documentTitle';
 import progressBar from './after/progressBar';
+import history from './after/history';
 
 export default (to) => {
     progressBar();
     documentTitle(to);
+    history(to);
 };