+++ resources/js/middleware/after/history.js
@@ -1,0 +1,7 @@
+import store from '../../store';
+
+export default (to) => {
+    if (to) {
+        store.commit('history/push', to);
+    }
+};