+++ resources/js/store/auth.js
@@ -1,6 +1,5 @@
 export const state = {
     isAuth: localStorage.getItem('isAuth') === 'true',
-    lastRoute: null,
 };
 
 export const mutations = {
@@ -12,5 +11,4 @@ export const mutations = {
         state.isAuth = false;
         localStorage.setItem('isAuth', false);
     },
-    setLastRoute: (state, route) => (state.lastRoute = route),
 };