+++ routes/web.php
@@ -1,5 +1,5 @@
 <?php
 
-Route::get('{any}', function () {
+Route::get('/{any}', function () {
     return view('laravel-enso/core::index');
 })->where('any', '.*');